Fix for grid file exports
The API was passing a string value of 'null' to PSaaS Builder
for the sub-scenario name. For some export types this was
being set as a valid sub-scenario name in the FGM which
caused PSaaS to miss the output as there was no matching
sub-scenario.

Make the API pass an empty string instead of null if no
sub-scenario name has been set to avoid this issue.